add App\Session::getUser

+++ view/forum/listPostsByTopic.php
@@ -16,8 +16,10 @@ $posts = $result["data"]["posts"];
                     <p>Créé par : <?= $post->getUser()->getId() ?> - <?= $post->getUser()->getNickname() ?></p>
                     <p>Créé le <?= $post->getDateCreation() ?></p>
                     <p><?= $post->getContent() ?></p>
-                    <a href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>" class="delete-button" onclick="return confirm('Êtes-vous sûr de vouloir supprimer ce post ?')">Supprimer</a>
-                    <a href="index.php?ctrl=forum&action=updatePost&id=<?= $post->getId() ?>" class="update-button" >Modifier</a>
+                    <?php if (App\Session::getUser()) : ?>
+                        <a href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>" class="delete-button" onclick="return confirm('Êtes-vous sûr de vouloir supprimer ce post ?')">Supprimer</a>
+                        <a href="index.php?ctrl=forum&action=updatePost&id=<?= $post->getId() ?>" class="update-button">Modifier</a>
+                    <?php endif; ?>
                 </div>
         <?php
             }
@@ -25,11 +27,13 @@ $posts = $result["data"]["posts"];
             echo "<p>Pas de post dans ce topic</p>";
         }
         ?>
-        <form action="index.php?ctrl=forum&action=addPost&id=<?= $topic->getId() ?>" method='POST' class="post-box">
-            <label for="post">Message :</label><br>
-            <textarea name="post" rows="4" placeholder="Entrez votre message" required></textarea><br>
-            <input type="submit" name="submit" value="Submit">
-        </form>
+        <?php if (App\Session::getUser()) : ?>
+            <form action="index.php?ctrl=forum&action=addPost&id=<?= $topic->getId() ?>" method='POST' class="post-box">
+                <label for="post">Message :</label><br>
+                <textarea name="post" rows="4" placeholder="Entrez votre message" required></textarea><br>
+                <input type="submit" name="submit" value="Submit">
+            </form>
+        <?php endif; ?>
     </div>
 
     <div class="topic-box">
